requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250706
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.13.4
blake3==1.0.5
bleach==6.3.0
blessed==1.21.0
blinker==1.9.0
blis==1.2.1
bm25s==0.2.0
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chonkie==1.0.6
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.6
compressed-tensors==0.9.2
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.3
cupy-cuda12x==13.5.1
cycler==0.12.1
cymem==2.0.11
datasets==3.5.1
decorator==5.2.1
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
faiss==1.8.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.2
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.60.0
frozenlist==1.6.0
fschat==0.2.36
fsspec==2025.3.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.8
hope==3.6.8
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
interegular==0.3.3
iocursor==0.1.4
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
lark==1.2.2
latex2mathml==3.78.0
llvmlite==0.43.0
lm-format-enforcer==0.10.12
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mistune==3.2.0
mkl==2024.2.2
mkl-service==2.4.0
modelscope==1.25.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
narwhals==1.37.1
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
ninja==1.11.1.4
nltk==3.9.1
numba==0.60.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.0
omegaconf==2.3.0
openai==2.6.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pdfminer.six==20250327
pdfplumber==0.11.6
Pebble==5.1.1
peft==0.15.2
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
preshed==3.0.9
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.5
pypdfium2==4.30.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rank-bm25==0.2.2
ray==2.45.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==5.1.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools-scm==9.0.1
sglang==0.4.6.post2
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
svgwrite==1.4.3
sympy==1.13.1
tbb==2021.13.1
tenacity==9.1.2
tensordict==0.5.0
thinc==8.3.4
threadpoolctl==3.6.0
thrift==0.21.0
tiktoken==0.9.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
vllm==0.8.1
wandb==0.19.10
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
zipp==3.21.0
